Iglu Core Circe: remove implicits.scala and rename instances.scala to implicits.scala (close #49)
